Fix genLaunchConfigFile
