+++ src/index.ts
@@ -8,9 +8,9 @@ export {
   findStructByType, findStructByName, isArrayType,
   arrayTypeAndSize, newCall, getNameByType, genLaunchConfigFile, subArrayType,
   isGenericType, parseGenericType,
-  readLaunchJson, getLowSPreimage, readBytes, parseAbiFromUnlockingScript, findConstStatic, findStatic, resolveConstValue,
+  readLaunchJson, getLowSPreimage, parseAbiFromUnlockingScript, findConstStatic, findStatic, resolveConstValue,
   arrayTypeAndSizeStr, toLiteralArrayType,
-  librarySign, structSign, resolveGenericType, canAssignProperty,
+  librarySign, structSign, resolveGenericType,
   buildTypeResolver, getStructDeclaration, getABIDeclaration, typeOfArg,
   compilerVersion, parseLiteral,
   isEmpty, JSONParser, getFullFilePath, path2uri, uri2path, md5, FunctionCall, stringToBytes, isScryptType, isSubBytes, toJSON,
