+++ src/internal.ts
@@ -8,3 +8,4 @@ export * from './findCompiler';
 export * from './partialSha256';
 export * from './typeCheck';
 export * from './builtins';
+export * from './launchConfig';
