+++ src/typeCheck.ts
@@ -1,6 +1,6 @@
-import { Argument, Arguments, isArrayType, isGenericType, LibraryEntity, ParamEntity, parseGenericType, StructEntity } from '.';
-import { subscript, toGenericType } from './internal';
-import { isBytes, ScryptType, StructObject, SupportedParamType, SymbolType, TypeResolver } from './scryptTypes';
+import { Argument, Arguments, isArrayType, LibraryEntity, ParamEntity, StructEntity } from '.';
+import { AliasEntity, StaticEntity, subscript, toGenericType } from './internal';
+import { isBytes, isScryptType, ScryptType, StructObject, SupportedParamType, SymbolType, TypeInfo, TypeResolver } from './scryptTypes';
 
 
 
@@ -240,7 +240,7 @@ export function checkSupportedParamType(arg: SupportedParamType, param: ParamEnt
   }
 }
 
-function hasGeneric(entity: StructEntity | LibraryEntity): boolean {
+export function hasGeneric(entity: StructEntity | LibraryEntity): boolean {
   return entity.genericTypes.length > 0;
 }
 
@@ -341,7 +341,7 @@ class GenericDeducer {
 export function deduceGenericStruct(param: ParamEntity, entity: StructEntity, resolver: TypeResolver): StructEntity | LibraryEntity | Error {
 
   if (!hasGeneric(entity)) {
-    return new Error(`struct ${entity.name} does not has any generic type`);
+    return entity;
   }
 
   const deducer = new GenericDeducer(entity, resolver);
@@ -356,7 +356,7 @@ export function deduceGenericStruct(param: ParamEntity, entity: StructEntity, re
 export function deduceGenericLibrary(param: ParamEntity, entity: LibraryEntity, resolver: TypeResolver): LibraryEntity | Error {
 
   if (!hasGeneric(entity)) {
-    return new Error(`library ${entity.name} does not has any generic type`);
+    return entity;
   }
 
   const deducer = new GenericDeducer(entity, resolver);
@@ -601,4 +601,185 @@ export function flatternArg(arg: Argument, resolver: TypeResolver, options: Flat
   }
 
   return args_;
+}
+
+export function findConstStatic(statics: StaticEntity[], name: string): StaticEntity | undefined {
+  return statics.find(s => {
+    return s.const === true && s.name === name;
+  });
+}
+export function findStatic(statics: StaticEntity[], name: string): StaticEntity | undefined {
+  return statics.find(s => {
+    return s.name === name;
+  });
+}
+
+
+function resolveAlias(alias: AliasEntity[], type: string): string {
+
+  const a = alias.find(a => {
+    return a.name === type;
+  });
+
+  if (a) {
+    return resolveAlias(alias, a.type);
+  }
+  return type;
+}
+
+
+export function resolveType(type: string, originTypes: Record<string, TypeInfo>, contract: string, statics: StaticEntity[], alias: AliasEntity[], librarys: LibraryEntity[]): TypeInfo {
+
+  type = resolveAlias(alias, type);
+
+  if (isArrayType(type)) {
+    const [elemTypeName, sizes] = arrayTypeAndSizeStr(type);
+    const elemTypeInfo = resolveType(elemTypeName, originTypes, contract, statics, alias, librarys);
+
+    if (isArrayType(elemTypeInfo.finalType)) {
+      const [elemTypeName_, sizes_] = arrayTypeAndSizeStr(elemTypeInfo.finalType);
+
+      const elemTypeInfo_ = resolveType(elemTypeName_, originTypes, contract, statics, alias, librarys);
+      return {
+        info: elemTypeInfo.info,
+        generic: elemTypeInfo.generic,
+        finalType: resolveConstStatic(contract, toLiteralArrayType(elemTypeInfo_.finalType, sizes.concat(sizes_)), statics),
+        symbolType: elemTypeInfo.symbolType
+      };
+    }
+
+    return {
+      info: elemTypeInfo.info,
+      generic: elemTypeInfo.generic,
+      finalType: resolveConstStatic(contract, toLiteralArrayType(elemTypeInfo.finalType, sizes), statics),
+      symbolType: elemTypeInfo.symbolType
+    };
+
+  } else if (isGenericType(type)) {
+    const [name, genericTypes] = parseGenericType(type);
+    const typeInfo = resolveType(name, originTypes, contract, statics, alias, librarys);
+    const gts = genericTypes.map(t => resolveType(t, originTypes, contract, statics, alias, librarys).finalType);
+
+    return {
+      info: typeInfo.info,
+      generic: true,
+      finalType: toGenericType(typeInfo.finalType, gts),
+      symbolType: typeInfo.symbolType
+    };
+  }
+
+  if (originTypes[type]) {
+    return originTypes[type];
+  } else if (isScryptType(type)) {
+    return {
+      finalType: type,
+      generic: false,
+      symbolType: SymbolType.ScryptType
+    };
+  } else {
+    return {
+      finalType: type,
+      generic: false,
+      symbolType: SymbolType.Unknown
+    };
+  }
+
+}
+
+
+
+function resolveConstStatic(contract: string, type: string, statics: StaticEntity[]): string {
+
+  if (isArrayType(type)) {
+    const [elemTypeName, arraySizes] = arrayTypeAndSizeStr(type);
+
+    const sizes = arraySizes.map(size => {
+      if (/^(\d)+$/.test(size)) {
+        return parseInt(size);
+      } else {
+        // size as a static const
+        const size_ = (size.indexOf('.') > 0) ? size : `${contract}.${size}`;
+        const value = findConstStatic(statics, size_);
+        if (!value) {
+          // Unable to solve when the subscript of the array is a function parameter, [CTC](https://scryptdoc.readthedocs.io/en/latest/ctc.html)
+          return size;
+        }
+        return value.value;
+      }
+    });
+
+    return toLiteralArrayType(elemTypeName, sizes);
+  }
+  return type;
+}
+
+
+/**
+ * check if a type is generic type
+ * @param type 
+ * @returns 
+ */
+export function isGenericType(type: string): boolean {
+  return /^([\w]+)<([\w,[\]\s<>]+)>$/.test(type);
+}
+
+/**
+ * 
+ * @param type eg. HashedMap<int,int>
+ * @param eg. ["HashedMap", ["int", "int"]}] An array generic types returned by @getGenericDeclaration
+ * @returns {"K": "int", "V": "int"}
+ */
+export function parseGenericType(type: string): [string, Array<string>] {
+
+  if (isGenericType(type)) {
+    const m = type.match(/([\w]+)<([\w,[\]<>\s]+)>$/);
+    if (m) {
+      const library = m[1];
+      const realTypes = [];
+      const brackets = [];
+      let tmpType = '';
+      for (let i = 0; i < m[2].length; i++) {
+        const ch = m[2].charAt(i);
+
+        if (ch === '<' || ch === '[') {
+          brackets.push(ch);
+        } else if (ch === '>' || ch === ']') {
+          brackets.pop();
+        } else if (ch === ',') {
+
+          if (brackets.length === 0) {
+            realTypes.push(tmpType.trim());
+            tmpType = '';
+            continue;
+          }
+        }
+        tmpType += ch;
+      }
+      realTypes.push(tmpType.trim());
+
+      return [library, realTypes];
+    }
+  }
+  throw new Error(`"${type}" is not generic type`);
+}
+
+
+
+
+export function resolveGenericType(genericTypeMap: Record<string, string>, type: string): string {
+  if (Object.keys(genericTypeMap).length > 0) {
+    if (isGenericType(type)) {
+      const [name, types] = parseGenericType(type);
+      return toGenericType(name, types.map(t => genericTypeMap[t] || t));
+    }
+
+    if (isArrayType(type)) {
+      const [elem, sizes] = arrayTypeAndSizeStr(type);
+      return toLiteralArrayType(elem, sizes.map(t => genericTypeMap[t] || t));
+    }
+
+    return genericTypeMap[type] || type;
+  }
+
+  return type;
 }
